+++ packages/avalanche-module/src/index.ts
@@ -1,2 +1,3 @@
 export * from './module';
 export * from './handlers/get-balances/typeguards';
+export * from './handlers/avalanche-send-transaction/typeguards';